llm-query-retrieval-hackrx/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
PyMuPDF==1.23.8
python-docx==1.1.0
openai==1.3.5
sentence-transformers==2.2.2
pinecone-client==2.2.4
pydantic==2.5.0
spacy==3.7.2
nltk==3.8.1
requests==2.31.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.3
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1
typing-extensions==4.8.0
